fix: ui revamp
