+++ src/routes/dashboard/guilds/+page.svelte
@@ -18,7 +18,7 @@
 	} from '$lib/ui/permMap';
 	import { CommonPermissionContext } from '../../../components/dashboard/permissions/commonPermissionContext';
 	import Message from '../../../components/Message.svelte';
-	import Guild from './Guild.svelte';
+	import Guild from './core/Guild.svelte';
 
 	let currentState = 'Loading dashboard data';
